// src/app/modules/core/models/auth.model.ts
export interface IUser {
  login: string;
  email: string;
  role: string;
}

export class User implements IUser {
  constructor(
    public login: string,
    public email: string,
    public role: string
  ) {}
}

export interface LoginData {
  login: string;
  password: string;
}

export interface ResetPasswordData {
  email: string;
}

export interface ChangePasswordData {
  password: string;
  uid: string;
}

export interface RegisterData {
  email: string;
  login: string;
  password: string;
}

export interface AuthResponse {
  timestamp: string;
  message: string;
  code: string;
}

export interface LoggedInResponse extends Omit<AuthResponse, 'message'> {
  message: boolean;
}